Moved common functions to AbstarctMinimap
