+++ Assets/Samples/MinimapFPS/UI/Minimap/SimpleStaticMinimap.prefab
@@ -59,4 +59,5 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   backgroundImage: {fileID: 21300000, guid: a93deafd9f0437c40ab6b2deb0eb8f54, type: 3}
-  MaskShape: {fileID: 21300000, guid: 53e0766c00bc26440968f31ce927cb2f, type: 3}
+  maskShape: {fileID: 21300000, guid: 53e0766c00bc26440968f31ce927cb2f, type: 3}
+  mapScale: {x: 2, y: 2}
