+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/AbstractMinimap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28a553a9f79ec5f49bb71fac03cc2c25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
