+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Centered/CenteredMinimap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98a32304b4e8fe1419e9210844b740c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
