+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Simple/SimpleStaticMinimap.cs
@@ -16,12 +16,10 @@
 // ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System.Collections.Generic;
 using com.nickmaltbie.MinimapTools.Icon;
 using com.nickmaltbie.MinimapTools.Minimap.MinimapBounds;
 using nickmaltbie.ScreenManager;
 using UnityEngine;
-using UnityEngine.UI;
 
 namespace com.nickmaltbie.MinimapTools.Minimap.Simple
 {
@@ -30,107 +28,27 @@ namespace com.nickmaltbie.MinimapTools.Minimap.Simple
     /// objects on top of a pre-rendered background image. Icons added to the
     /// map will have their position updated each frame to follow the minimap.
     /// </summary>
-    public class SimpleStaticMinimap : MonoBehaviour, IMinimap, IScreenComponent
+    public class SimpleStaticMinimap : AbstractMinimap, IScreenComponent
     {
         /// <summary>
-        /// Default bounds for the minimap.
+        /// Bounds source for the minimap.
         /// </summary>
-        private static readonly Bounds defaultBounds = new Bounds(Vector3.zero, Vector3.one * 10);
-
-        /// <summary>
-        /// Background image for the minimap.
-        /// </summary>
-        [SerializeField]
-        [Tooltip("Background image for the minimap.")]
-        private Sprite backgroundImage;
-
-        /// <summary>
-        /// Shape of the minimap mask.
-        /// </summary>
-        [SerializeField]
-        [Tooltip("Shape of the minimap mask.")]
-        private Sprite MaskShape;
-
-        /// <summary>
-        /// Source of bounds for the minimap.
-        /// </summary>
-        private IBoundsSource source;
-
-        /// <summary>
-        /// Collection of icons contained in this minimap.
-        /// </summary>
-        private Dictionary<IMinimapIcon, GameObject> icons = new Dictionary<IMinimapIcon, GameObject>();
-
-        /// <summary>
-        /// Gets the world bounds for this simple minimap.
-        /// </summary>
-        private Bounds WorldBounds => source?.GetBounds() ?? defaultBounds;
-
-        public void LateUpdate()
-        {
-            foreach (IMinimapIcon icon in icons.Keys)
-            {
-                UpdateMinimapIconPosition(icon);
-            }
-        }
-
-        public void Start()
-        {
-            foreach (IMinimapIcon icon in GameObject.FindObjectsOfType<SpriteIcon>())
-            {
-                AddIcon(icon);
-            }
-        }
-
-        public void Awake()
-        {
-            _ = gameObject.AddComponent<Mask>();
-            Image maskImage = gameObject.AddComponent<Image>();
-            maskImage.sprite = MaskShape;
-
-            var background = new GameObject();
-            background.name = "Background";
-            background.transform.SetParent(transform);
-
-            RectTransform backgroundTr = background.AddComponent<RectTransform>();
-            backgroundTr.anchorMin = Vector2.zero;
-            backgroundTr.anchorMax = Vector2.one;
-            backgroundTr.pivot = Vector3.zero;
-            backgroundTr.sizeDelta = Vector2.one;
-            backgroundTr.anchoredPosition = Vector2.zero;
-
-            Image image = background.AddComponent<Image>();
-            image.sprite = backgroundImage;
-        }
+        private IBoundsSource boundsSource;
 
         /// <inheritdoc/>
-        public bool AddIcon(IMinimapIcon minimapIcon)
-        {
-            if (icons.ContainsKey(minimapIcon))
-            {
-                return false;
-            }
-
-            GameObject iconGo = minimapIcon.CreateIcon(GetComponent<RectTransform>());
-            iconGo.transform.SetParent(transform);
-            iconGo.name = "icon";
-
-            icons.Add(minimapIcon, iconGo);
-            UpdateMinimapIconPosition(minimapIcon);
-            return true;
-        }
+        protected override IBoundsSource Source => boundsSource;
 
         /// <inheritdoc/>
-        public bool InMap(Vector3 worldSpace)
-        {
-            return WorldBounds.Contains(worldSpace);
-        }
+        protected override Vector2 MapOffset => Vector2.zero;
+
+        /// <inheritdoc/>
+        protected override Vector2 MapScale => Vector2.one;
 
         /// <inheritdoc/>
         public void OnScreenLoaded()
         {
             // Update bounds source based on source in scene
-            source ??= GameObject.FindObjectOfType<BoxBoundsSource>() as IBoundsSource;
+            boundsSource ??= GameObject.FindObjectOfType<BoxBoundsSource>() as IBoundsSource;
         }
 
         /// <inheritdoc/>
@@ -139,45 +57,12 @@ namespace com.nickmaltbie.MinimapTools.Minimap.Simple
 
         }
 
-        /// <inheritdoc/>
-        public bool RemoveIcon(IMinimapIcon minimapIcon)
+        public void Start()
         {
-            if (icons.TryGetValue(minimapIcon, out GameObject go))
+            foreach (IMinimapIcon icon in GameObject.FindObjectsOfType<SpriteIcon>())
             {
-                GameObject.DestroyImmediate(go);
+                AddIcon(icon);
             }
-
-            return icons.Remove(minimapIcon);
-        }
-
-        /// <summary>
-        /// UPdate the position of a minimap icon based on its current position.
-        /// </summary>
-        /// <param name="icon">icon to update position of.</param>
-        private void UpdateMinimapIconPosition(IMinimapIcon icon)
-        {
-            if (icons.TryGetValue(icon, out GameObject iconGo))
-            {
-                Vector3 relativePosition = GetMinimapPosition(icon.GetWorldSpace());
-                RectTransform rectTransform = iconGo.GetComponent<RectTransform>();
-                rectTransform.anchorMax = relativePosition;
-                rectTransform.anchorMin = relativePosition;
-                rectTransform.anchoredPosition = Vector2.zero;
-                rectTransform.rotation = Quaternion.Euler(0, 0, -icon.GetIconRotation().eulerAngles.y);
-            }
-        }
-
-        /// <summary>
-        /// Translates a position from world space to normalized minimap space.
-        /// </summary>
-        /// <param name="worldPosition">Position of the object in world space.</param>
-        /// <returns>Normalized minimap position, will scale positions within
-        /// the minimap to between (0,0) and (1,1).</returns>
-        private Vector2 GetMinimapPosition(Vector3 worldPosition)
-        {
-            Vector3 relativePosition = worldPosition - WorldBounds.min;
-            var normalizedPosition = new Vector2(relativePosition.x / WorldBounds.size.x, relativePosition.z / WorldBounds.size.z);
-            return normalizedPosition;
         }
     }
 }
