--- Packages/com.nickmaltbie.MinimapTools/Scripts/Simple.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9a0815e37218b8346851099dd19e4b05
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
